+++ src/components/DesignSystemButton/DesignSystemButton.jsx
@@ -1,0 +1,6 @@
+import { Button } from "@RaspberryPiFoundation/design-system-react";
+import "../../assets/stylesheets/DesignSystemButton.scss";
+
+const DesignSystemButton = (props) => <Button {...props} />;
+
+export default DesignSystemButton;