+++ src/components/RunButton/RunBar.jsx
@@ -1,0 +1,14 @@
+import React from "react";
+import RunnerControls from "./RunnerControls";
+
+import "../../assets/stylesheets/RunBar.scss";
+
+const RunBar = ({ embedded = false }) => {
+  return (
+    <div className="run-bar">
+      <RunnerControls embedded={embedded} />
+    </div>
+  );
+};
+
+export default RunBar;