+++ CHANGELOG.md
@@ -8,9 +8,21 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [0.19.1] - 2023-10-04
 
+### Changed
+
+- Performance enhancements on app routes & sense hat transitions (#686)
+- Web component styling (#687)
+- Convert React components to JSX files (#688)
+- Moved `svg`s to `/src/assets/icons` (#692)
+- Moved `scss` files to `/src/stylesheets` (#692)
+- Moved slices and reducers to `/src/redux` (#692)
+- Moved loaders to `/src/containers` (#692)
+
+
 ### Fixed
 
 - Min-height on Sense HAT model to fix tablet-view (#658)
+- Tidied up stylesheets (#684)
 
 ## [0.19.0] - 2023-09-25